fileを合併
